+++ apps/pie-docs/src/componentStatusData.js
@@ -200,45 +200,6 @@ const rows = [
         }
     ],
     [
-        { componentName: 'Banner' },
-        {
-            resource: resourceTypes.COMPONENT,
-            link: 'https://www.figma.com/file/R2rBfzJP0hG0MZorq6FLZ1/%5BCore%5D-Components-%E2%9A%AA%EF%B8%8F-%5BPIE-2.0%5D?type=design&node-id=877-2984&mode=design',
-            status: statusTypes.STABLE,
-        },
-        {
-            resource: resourceTypes.DOCUMENTATION,
-            link: 'https://www.figma.com/file/j1YKygEyhqZ6zKVxcHapn5/%5BCore%5D-Component-Documentation-%E2%9A%AA%EF%B8%8F-%5BPIE-2.0%5D?type=design&node-id=877-2984&mode=design',
-            status: statusTypes.STABLE,
-        },
-        {
-            resource: resourceTypes.WEB_COMPONENTS,
-            status: statusTypes.PLANNED,
-        },
-        {
-            resource: resourceTypes.VUE,
-            link: 'https://vue.pie.design/?path=/story/components-molecules--alert-component',
-            status: statusTypes.STABLE,
-        },
-        {
-            resource: resourceTypes.REACT,
-            link: 'https://snacks.takeaway.com/portal/components/notification/',
-            status: statusTypes.STABLE,
-        },
-        {
-            resource: resourceTypes.IOS_JUST_UI,
-            status: statusTypes.STABLE,
-        },
-        {
-            resource: resourceTypes.IOS_PIE,
-            status: statusTypes.STABLE,
-        },
-        {
-            resource: resourceTypes.ANDROID_PIE,
-            status: statusTypes.STABLE,
-        }
-    ],
-    [
         { componentName: 'Bottom Sheet' },
         {
             resource: resourceTypes.COMPONENT,