+++ apps/examples/wc-next10/CHANGELOG.md
@@ -1,5 +1,26 @@
 # wc-next10
 
+## 0.6.21
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc@0.5.35
+
+## 0.6.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc@0.5.34
+
+## 0.6.19
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc@0.5.33
+
 ## 0.6.18
 
 ### Patch Changes